Move around some modules, add comments.
